requirements.txt
Flask==3.0.3
librosa==0.10.2
numpy==1.26.4
transformers==4.44.2
torch>=2.6.0
gTTS==2.5.3
python-Levenshtein==0.25.1
dtw-python==1.5.1
safetensors==0.4.5
huggingface_hub==0.25.2
soundfile==0.12.1 
pydub==0.25.1
jiwer==3.0.4